Pull image regardless of whether it exists locally if `tag=latest`
... and pull policy is `missing` or `if_not_present`
