+++ pkg/e2e/fixtures/compose-pull/image-present-locally/docker-compose.yaml
@@ -3,3 +3,7 @@ services:
     image: alpine:3.13.12
     pull_policy: missing
     command: top
+  latest:
+    image: alpine:latest
+    pull_policy: missing
+    command: top
